`public_api`: Introduce private `mod public_item`
When working on improved support for rendering `impl`s, I found
item_iterator.rs to be inconveniently big. So move `PublicItem` and
friends out to a new `mod public_item`.
